+++ src/corpus/config/CorpusConfigCustomEdit.vue
@@ -8,7 +8,6 @@ import { isError } from "es-toolkit";
 import useCorpusIdParam from "../corpusIdParam.composable";
 import useConfig from "./config.composable";
 import CorpusConfigCustomHelp from "./CorpusConfigCustomHelp.vue";
-import { useAuth } from "@/auth/auth.composable";
 import useMessenger from "@/message/messenger.composable";
 import LayoutBox from "@/components/LayoutBox.vue";
 import HelpBox from "@/components/HelpBox.vue";
@@ -21,7 +20,6 @@ const ajv = new Ajv2020();
 const schemaValidate = ajv.compile(schema);
 
 const corpusId = useCorpusIdParam();
-const { requireAuthentication } = useAuth();
 const { config, uploadConfigRaw } = useConfig(corpusId);
 const { alertError } = useMessenger();
 const { t } = useI18n();
@@ -29,8 +27,6 @@ const { t } = useI18n();
 const input = ref(config.value || "");
 const error = ref<{ hint: string; details?: string }>();
 const showErrorDetails = ref<boolean>(false);
-
-requireAuthentication();
 
 function validate() {
   error.value = undefined;