+++ src/resource/ResourceRedirectView.vue
@@ -1,22 +1,20 @@
 <script setup lang="ts">
 import { useRouter } from "vue-router";
 import useResourceIdParam from "@/resource/resourceIdParam.composable";
-import { useAuth } from "@/auth/auth.composable";
 import useResources from "@/library/resources.composable";
+import SpinIndicator from "@/spin/SpinIndicator.vue";
 
 const router = useRouter();
-const { requireAuthentication } = useAuth();
 const resourceId = useResourceIdParam();
 const { loadResource } = useResources();
 
-requireAuthentication(async () => {
-  // Load resource and immediately redirect according to its type.
-  const resource = await loadResource(resourceId);
-  if (!resource) return;
-  router.push(`/library/${resource.type}/${resourceId}`);
+// Load resource and immediately redirect according to its type.
+loadResource(resourceId).then((resource) => {
+  if (resource) router.push(`/library/${resource.type}/${resourceId}`);
+  else router.push("/library");
 });
 </script>
 
 <template>
-  <div />
+  <SpinIndicator />
 </template>