+++ CHANGELOG.md
@@ -21,6 +21,7 @@ As this project is a user-facing application, the places in the semantic version
 
 - Use `/jwt?version=2` to avoid console error when not authenticated
 - Use corpusState only for guiding the user [#149](https://github.com/spraakbanken/mink-frontend/issues/149)
+- Mark routes that require authentication on router level (`meta: {protected: true}`)
 
 ### Fixed